app/requirements.txt
django==4.0.4
django-split-settings==1.2.0
elastic-transport==8.4.0
elasticsearch==8.5.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pydantic==1.10.2
uwsgi==2.0.20
tqdm==4.64.1